Adds jest and @testing-library/react
