+++ jest.setup.js
@@ -1,0 +1,3 @@
+import recurly from 'recurly.js';
+
+global.recurly = recurly;
